updated install script with checks

+++ sssd-install.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-
 ##check if root
-
+if [[ $EUID -ne 0 ]]; then
+   echo "This script must be run as root, exiting."
+   exit 1
+fi
 
 #fix /etc/skel
 mkdir /etc/skel/.ssh
@@ -15,18 +17,23 @@ groupadd -g 100000 wsusers
 echo "AllowGroups users wsusers mkijowski" >> /etc/ssh/sshd_config
 
 #may not be necessary if we fix sssd, WSU thinks tcsh is default sh
-ln -s /bin/bash /bin/tcsh
+ln -sfb /bin/bash /bin/tcsh
 
 #check if dir exists
-mkdir /etc/sssd
-#touch /etc/sssd/sssd.conf
-cp /home/conf/sssd.conf /etc/sssd/sssd.conf
+mkdir -p /etc/sssd
+
+if [ -f /home/conf/sssd.conf ]; then
+    cp /home/conf/sssd.conf /etc/sssd/sssd.conf
+else
+    echo "No sssd.conf found, creating blank file for permissions"
+    touch /etc/sssd/sssd.conf
+fi
+
 chown root:root /etc/sssd/sssd.conf
 chmod 700 /etc/sssd/sssd.conf
 
 apt update && apt install -y sssd
 
-#probably should make sure common-session layout is the same in 16.04 and 18.04
 cat ./common-session.ed | ed - /etc/pam.d/common-session
 
 service sssd restart
